modify(entity): 금액 관련 필드 long 타입으로 통일

+++ src/main/java/com/nhnacademy/back/account/memberrank/domain/entity/MemberRank.java
@@ -28,6 +28,6 @@ public class MemberRank {
 	private int memberRankTierBonusRate;
 
 	@Column(nullable = false)
-	private int memberRankRequireAmount;
+	private long memberRankRequireAmount;
 
 }
