+++ src/main/java/com/nhnacademy/back/coupon/couponpolicy/domain/entity/CouponPolicy.java
@@ -23,9 +23,9 @@ public class CouponPolicy {
 	@Column(nullable = false, columnDefinition = "integer DEFAULT 0")
 	private int couponPolicyMinimum = 0;
 
-	private Integer couponPolicyMaximumAmount;
+	private Long couponPolicyMaximumAmount;
 
-	private Integer couponPolicySalePrice;
+	private Long couponPolicySalePrice;
 
 	private Integer couponPolicyDiscountRate;
